PokerBestHandExtension/src/test/java/com/amaap/pokerbesthandextension/domain/service/StraightRankEvaluatorTest.java
package com.amaap.pokerbesthandextension.domain.service;

import com.amaap.pokerbesthandextension.domain.model.valueobject.HandRank;
import com.amaap.pokerbesthandextension.repository.dto.CardDto;
import com.amaap.pokerbesthandextension.repository.dto.builder.CardBuilder;
import com.amaap.pokerbesthandextension.repository.dto.exception.InvalidCardCodeException;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class StraightRankEvaluatorTest {
    StraightRankEvaluator straightRankEvaluator = new StraightRankEvaluator();
    @Test
    void shouldBeAbleToEvaluateStraightRank() throws InvalidCardCodeException {
        // arrange
        List<CardDto> cards = CardBuilder.getCardsForStraightRank();
        HandRank expected = HandRank.STRAIGHT;

        // act
        HandRank actual = straightRankEvaluator.getRank(cards);

        // assert
        assertEquals(expected,actual);
    }

    @Test
    void shouldBeAbleToGetNullWhenRankIsNotStraight() throws InvalidCardCodeException {
        // arrange
        List<CardDto> cards = CardBuilder.getFiveCards();

        // act
        HandRank rank = straightRankEvaluator.getRank(cards);

        // assert
        assertNull(rank);
    }
}